+++ anytime-fitness/src/Components/Class.js
@@ -1,0 +1,8 @@
+import React from "react";
+
+export default function Class(props) {
+
+    return (
+        //class construction here
+    )
+}
