+++ anytime-fitness/src/Components/UserDash.js
@@ -1,0 +1,16 @@
+import React from "react";
+import { useHistory } from "react-router-dom"; 
+
+export default function UserDash () {
+    const history = useHistory()
+
+    const routeToUserHome = () => {
+        history.push('/userhome');
+    }
+
+    return (
+        <div className='home container'>
+           
+        </div>
+    )
+}
